feat: nested cli sub-commands for plots
