+++ cakes-results/python/cakes_results/__init__.py
@@ -1,1 +1,3 @@
 """Plots and analysis of the results of the Cakes project."""
+
+from . import scaling  # noqa: F401
