+++ cakes-results/python/cakes_results/scaling/__init__.py
@@ -1,0 +1,5 @@
+"""Plots and analysis of the search scaling results for Cakes."""
+
+from . import reports  # noqa: F401
+from .plots import create_plots  # noqa: F401
+from .app import app  # noqa: F401
