+++ cakes-results/python/cakes_results/scaling/plots.py
@@ -1,38 +1,9 @@
-"""Plots for the scaling results of the Cakes scaling benchmarks."""
-
 import enum
-import json
 import pathlib
-import typing
 
 import matplotlib.pyplot as plt
-import pandas
-import pydantic
 
-
-class Report(pydantic.BaseModel):
-    """Report of the scaling results of the Cakes search."""
-
-    dataset: str
-    metric: str
-    base_cardinality: int
-    dimensionality: int
-    num_queries: int
-    error_rate: float
-    ks: list[int]
-    csv_path: pathlib.Path = pathlib.Path(".").resolve()
-
-    @staticmethod
-    def from_json(json_path: pathlib.Path) -> "Report":
-        """Load the report from a JSON file."""
-        with json_path.open("r") as json_file:
-            contents: dict[str, typing.Any] = json.load(json_file)
-            contents["csv_path"] = json_path.parent.joinpath(contents.pop("csv_name"))
-            return Report(**contents)
-
-    def to_pandas(self) -> pandas.DataFrame:
-        """Read the CSV file into a pandas DataFrame."""
-        return pandas.read_csv(self.csv_path)
+from .reports import Report
 
 
 class Markers(str, enum.Enum):
@@ -71,7 +42,7 @@ class Markers(str, enum.Enum):
         return m
 
 
-def plot_throughput(
+def create_plots(
     json_path: pathlib.Path, make_title: bool, output_dir: pathlib.Path
 ) -> bool:
     """Plot the throughput of the Cakes search."""
